move string extension
