+++ CodeEdit/Info.plist
@@ -45,6 +45,6 @@
 		</dict>
 	</array>
 	<key>GitHash</key>
-	<string>601e98c51c1347f0f24eea3a87416a8c10bbc5d4</string>
+	<string>cde35bdfc4817417c9ea0f964ca9da086aeb7301</string>
 </dict>
 </plist>
